@@ -1,40 +1,56 @@
 {
 	"name": "chat-transmitter-bot",
-	"version": "1.0.0",
+	"version": "2.0.0",
 	"description": "",
 	"main": "build/main.js",
 	"scripts": {
 		"clean": "rimraf build/",
 		"build": "tsc",
 		"start": "node build/main.js",
-		"watch": "concurrently \"tsc -w --project .\" \"npm run nodemon\"",
-		"nodemon": "nodemon -q -w build build/main.js",
-		"lint": "tslint --project .",
-		"typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js",
+		"watch": "tsc-watch -p . --onSuccess \"node build/main.js\"",
+		"lint": "eslint src/",
+		"typeorm": "typeorm-ts-node-esm",
+		"migration:create": "typeorm-ts-node-esm migration:create",
+		"migration:generate": "typeorm-ts-node-esm migration:generate -d build/dataSource.js",
+		"registercommands": "node build/registerCommands.js",
+		"unregistercommands": "node build/unregisterCommands.js",
 		"test": "echo \"Error: no test specified\" && exit 1"
 	},
 	"author": "r-o-b-o-t-o",
 	"license": "MIT",
 	"devDependencies": {
-		"@typeit/discord": "^4.0.10",
-		"@types/node": "^16.10.3",
-		"@types/ws": "^8.2.0",
-		"concurrently": "^6.3.0",
-		"nodemon": "^2.0.13",
+		"@types/express": "^4.17.13",
+		"@types/fs-extra": "^9.0.13",
+		"@types/node": "^18.7.2",
+		"@types/node-cron": "^3.0.2",
+		"@types/ws": "^8.5.3",
+		"@typescript-eslint/eslint-plugin": "^5.33.0",
+		"@typescript-eslint/parser": "^5.33.0",
+		"concurrently": "^7.3.0",
+		"eslint": "^8.21.0",
 		"rimraf": "^3.0.2",
-		"ts-node-dev": "^1.1.8",
-		"tslint": "^6.1.3",
-		"typescript": "^4.4.3"
+		"ts-node": "^10.9.1",
+		"tsc-watch": "^5.0.3",
+		"typescript": "^4.7.4"
 	},
 	"dependencies": {
-		"discord.js": "^12.5.3",
-		"nanoid": "^3.1.29",
-		"node-fetch": "^2.6.5",
-		"query-string": "^7.0.1",
+		"@discordjs/builders": "^1.1.0",
+		"@discordjs/rest": "^1.0.1",
+		"csv": "^6.2.0",
+		"discord-api-types": "^0.37.2",
+		"discord.js": "^14.2.0",
+		"express": "^4.18.1",
+		"fs-extra": "^10.1.0",
+		"glob-promise": "^4.2.2",
+		"nanoid": "^3.3.4",
+		"node-cron": "^3.0.1",
+		"node-fetch": "^2.6.7",
+		"query-string": "^7.1.1",
 		"reflect-metadata": "^0.1.13",
-		"sqlite3": "^5.0.2",
-		"typeorm": "^0.2.38",
-		"typeorm-naming-strategies": "^2.0.0",
-		"ws": "^8.2.3"
+		"source-map-support": "^0.5.21",
+		"sqlite3": "^5.0.11",
+		"typeorm": "^0.3.7",
+		"typeorm-naming-strategies": "^4.1.0",
+		"ws": "^8.8.1"
 	}
 }
