@@ -1,30 +1,34 @@
-import * as fs from "fs";
-
-const fsAsync = fs.promises;
+import fs from "fs-extra";
 
 export class Config {
-	public listenPort: number;
+	public wsPort: number;
+	public httpHost: string;
+	public httpPort: number;
 	public secretKey: string;
+	public discordClientId: string;
 	public discordToken: string;
 	public useClassEmoji: boolean;
 	public classEmojis: string[];
 	public useRaceEmoji: boolean;
-	public raceEmojis: string[];
+	public raceEmojis: string[][];
 	public useRaidMarkerEmoji: boolean;
 	public raidMarkerEmojis: string[];
+	public confirmEmojis: string[];
 	public filterAtHere: boolean;
 	public filterAtEveryone: boolean;
 	public apiBaseUrl: string;
 	public ignoreGuilds: string[];
+	public queryResultsKeepDuration: number;
+	public anticheatReportThrottleDuration: number;
 
-	private static checkedMissingField: boolean = false;
+	private static checkedMissingField: boolean;
 
 	public static async load(): Promise<Config> {
-		const json: Buffer = await fsAsync.readFile("config.json");
+		const json: Buffer = await fs.readFile("config.json");
 		const config = JSON.parse(json.toString()) as Config;
 
 		if (!Config.checkedMissingField) {
-			const defaultConfigJson = await fsAsync.readFile("config.default.json");
+			const defaultConfigJson = await fs.readFile("config.default.json");
 			const defaultConfig = JSON.parse(defaultConfigJson.toString());
 			Config.checkAllMissingFields(config, defaultConfig);
 			Config.checkedMissingField = true;
@@ -33,7 +37,7 @@ export class Config {
 		return config;
 	}
 
-	private static checkAllMissingFields(obj: object, model: object, parentName: string = "") {
+	private static checkAllMissingFields(obj: object, model: object, parentName = "") {
 		const missing = Config.hasMissingFields(obj, model);
 		if (parentName !== "") {
 			parentName += ".";
@@ -42,19 +46,19 @@ export class Config {
 			console.warn(`Field ${parentName}${field} is missing in config.json!`);
 		}
 		for (const key in model) {
-			if (typeof model[key] === "object" && obj.hasOwnProperty(key)) {
+			if (typeof model[key] === "object" && Object.prototype.hasOwnProperty.call(obj, key)) {
 				Config.checkAllMissingFields(obj[key], model[key], parentName + key);
 			}
 		}
 	}
 
 	private static hasMissingFields(obj: object, model: object): string[] {
-		const missing = [];
+		const missing: string[] = [];
 		for (const key in model) {
-			if (!obj.hasOwnProperty(key)) {
+			if (!Object.prototype.hasOwnProperty.call(obj, key)) {
 				missing.push(key);
 			}
 		}
 		return missing;
 	}
-};
+}
