@@ -1,63 +1,73 @@
 import { Config } from "./Config";
 import { IpBan } from "./entity/IpBan";
+import { EDatabase } from "./entity/Query";
 import { IEvent } from "./controller/Event";
-import { ChatController } from "./controller/ChatController";
-import { CommandController } from "./controller/CommandController";
 import { WebSocketWrapper } from "./WebSocketWrapper";
+import { ChatController } from "./controller/ChatController";
+import { QueryController } from "./controller/QueryController";
 import { WebSocketServerWrapper } from "./WebSocketServerWrapper";
+import { CommandController } from "./controller/CommandController";
+import { AnticheatController } from "./controller/AnticheatController";
 
 export class WebSocketManager {
 	public static instance: WebSocketManager;
 
-	private connection: WebSocketWrapper;
+	private config: Config;
+	private connection: WebSocketWrapper | null;
 
-	public constructor() {
+	public constructor(config: Config) {
 		WebSocketManager.instance = this;
+		this.config = config;
 	}
 
-	public start(config: Config): void {
+	public run(): void {
 		const keyInvalidCharacters = ["&", "#", "\""];
 		for (const char of keyInvalidCharacters) {
-			if (config.secretKey.includes(char)) {
+			if (this.config.secretKey.includes(char)) {
 				throw new Error(`"secretKey" configuration field contains invalid character '${char}'`);
 			}
 		}
 
 		const wss = new WebSocketServerWrapper({
 			host: "0.0.0.0",
-			port: config.listenPort,
+			port: this.config.wsPort,
+			maxPayload: 50 * 1024 * 1024, // 50 MB
 		}, () => {
-			console.log(`WebSocket server is listening on 0.0.0.0:${config.listenPort}.`);
+			console.log(`WebSocket server is listening on 0.0.0.0:${this.config.wsPort}.`);
 		});
 
-		const controllers = [ChatController, CommandController].map(controller => new controller());
-		const handlers: { [key: string]: (data: any) => void } = { };
+		const controllers = [ChatController, CommandController, QueryController, AnticheatController].map(controller => new controller());
+		const handlers: { [key: string]: (data) => void } = {};
 		for (const controller of controllers) {
 			const events: IEvent[] = Reflect.getMetadata("events", controller.constructor);
 
 			if (events) {
 				for (const event of events) {
-					handlers[event.eventName] = (data: any) => controller[event.methodName](data);
+					handlers[event.eventName] = (data) => controller[event.methodName](data);
 				}
 			}
 		}
 
 		wss.onConnection(async (ws, connData) => {
+			if (!connData) {
+				return;
+			}
 			const ipBan = await IpBan.find(ws.remoteAddress);
 			if (ipBan?.banned) {
 				ws.close(4001, "Access denied");
 				return;
 			}
 
-			if (connData.key === undefined || (connData.key as string) !== config.secretKey) {
+			if (connData.key === undefined || (connData.key as string) !== this.config.secretKey) {
 				ws.close(4000, "Incorrect key");
 				IpBan.addConnectionAttempt(ws.remoteAddress);
 				return;
 			}
 
 			this.connection = ws;
+			console.log("Accepted WebSocket connection from " + ws.remoteAddress);
 
-			ws.on("message", (name: string, data: any) => {
+			ws.on("message", (name: string, data) => {
 				handlers[name]?.(data);
 			});
 			ws.on("close", () => {
@@ -77,4 +87,17 @@ export class WebSocketManager {
 		});
 		return true;
 	}
-};
+
+	public sendQuery(id: string, query: string, database: EDatabase): boolean {
+		if (!this.connection) {
+			return false;
+		}
+
+		this.connection.send("query", {
+			id,
+			query,
+			database,
+		});
+		return true;
+	}
+}
