@@ -1,4 +1,6 @@
-import { Entity, Column, PrimaryColumn, getManager, Raw } from "typeorm";
+import { Entity, Column, PrimaryColumn, Raw } from "typeorm";
+
+import { db } from "../dataSource";
 
 @Entity()
 export class Area {
@@ -8,15 +10,19 @@ export class Area {
 	@Column()
 	name: string;
 
+	public static async findAll(): Promise<Area[]> {
+		return await db.find(Area);
+	}
+
 	public static async find(name: string): Promise<Area[]> {
-		return await getManager().find(Area, {
+		return await db.find(Area, {
 			where: {
 				name: Raw((column: string) => `LOWER(${column}) LIKE LOWER('%${name}%')`),
 			},
 		});
 	}
 
-	public static async findById(id: number): Promise<Area> {
-		return await getManager().findOne(Area, id);
+	public static async findById(id: number): Promise<Area | null> {
+		return await db.findOneBy(Area, { id });
 	}
-};
+}
