@@ -1,4 +1,9 @@
-import { Entity, PrimaryColumn, Column, getManager, OneToMany } from "typeorm";
+import { GuildMember } from "discord.js";
+import { Entity, PrimaryColumn, Column, OneToMany } from "typeorm";
+
+import { Bot } from "../Bot";
+import { Query } from "./Query";
+import { db } from "../dataSource";
 import { Channel } from "./Channel";
 import { AdminRole } from "./AdminRole";
 import { ZoneChannel } from "./ZoneChannel";
@@ -8,49 +13,60 @@ export class Guild {
 	@PrimaryColumn()
 	discordId: string;
 
-	@Column({ "nullable": true })
+	@Column({ nullable: true })
 	miscChannel: string;
 
-	@OneToMany(type => Channel, chan => chan.guild)
+	@Column({ nullable: true })
+	anticheatReportsChannel: string;
+
+	@OneToMany(_type => Channel, chan => chan.guild)
 	channels: Channel[];
 
-	@OneToMany(type => ZoneChannel, chan => chan.guild)
+	@OneToMany(_type => ZoneChannel, chan => chan.guild)
 	zoneChannels: ZoneChannel[];
 
-	@OneToMany(type => AdminRole, role => role.guild)
+	@OneToMany(_type => AdminRole, role => role.guild)
 	adminRoles: AdminRole[];
 
-	public static async findOrCreate(discordId: string): Promise<Guild> {
-		let guild: Guild = await this.find(discordId);
+	@OneToMany(_type => Query, query => query.guild)
+	queries: Query[];
 
-		if (guild === undefined) {
+	public static async findOrCreate(discordId: string): Promise<Guild> {
+		let guild = await Guild.find(discordId);
+
+		if (!guild) {
 			guild = new Guild();
 			guild.discordId = discordId;
-			await getManager().save(guild);
+			await guild.save();
+			guild = await Guild.find(discordId);
+			if (!guild) {
+				throw new Error(`Could not find guild ${discordId} after its creation`);
+			}
 		}
 
 		return guild;
 	}
 
-	public static async find(discordId: string): Promise<Guild> {
-		return await getManager().findOne(Guild, discordId, {
-			relations: ["channels", "zoneChannels", "adminRoles"],
+	public static async find(discordId: string): Promise<Guild | null> {
+		return await db.findOne(Guild, {
+			where: { discordId },
+			relations: ["channels", "zoneChannels", "adminRoles", "adminRoles.guild", "queries"],
 		});
 	}
 
-	public getChannel(channelName: string): Channel {
+	public getChannel(channelName: string): Channel | undefined {
 		return this.channels.find((c) => c.name === channelName);
 	}
 
-	public getZoneChannel(zoneName: string): Channel {
+	public getZoneChannel(zoneName: string): Channel | undefined {
 		return this.zoneChannels.find((c) => c.name === zoneName);
 	}
 
 	public async addChannel(discordId: string, channelName: string): Promise<Channel> {
-		let channel: Channel = this.getChannel(channelName);
-		if (channel !== undefined) {
+		let channel = this.getChannel(channelName);
+		if (channel) {
 			this.channels.splice(this.channels.indexOf(channel), 1);
-			await getManager().remove(Channel, channel);
+			await db.remove(Channel, channel);
 		}
 
 		channel = new Channel();
@@ -59,15 +75,15 @@ export class Guild {
 		channel.guild = this;
 		this.channels.push(channel);
 
-		await getManager().save(channel);
+		await db.save(channel);
 		return channel;
 	}
 
 	public async addZoneChannel(discordId: string, zoneName: string): Promise<ZoneChannel> {
-		let channel: ZoneChannel = this.getZoneChannel(zoneName);
-		if (channel !== undefined) {
+		let channel = this.getZoneChannel(zoneName);
+		if (channel) {
 			this.zoneChannels.splice(this.zoneChannels.indexOf(channel), 1);
-			await getManager().remove(ZoneChannel, channel);
+			await db.remove(ZoneChannel, channel);
 		}
 
 		channel = new ZoneChannel();
@@ -76,13 +92,13 @@ export class Guild {
 		channel.guild = this;
 		this.zoneChannels.push(channel);
 
-		await getManager().save(channel);
+		await db.save(channel);
 		return channel;
 	}
 
 	public async addAdminRole(discordId: string): Promise<AdminRole> {
-		let role: AdminRole = this.adminRoles.find(r => r.discordId === discordId);
-		if (role !== undefined) {
+		let role = this.adminRoles.find(r => r.discordId === discordId);
+		if (role) {
 			return role;
 		}
 
@@ -91,16 +107,35 @@ export class Guild {
 		role.guild = this;
 		this.adminRoles.push(role);
 
-		await getManager().save(role);
+		await db.save(role);
 		return role;
 	}
 
-	public async removeAdminRole(discordId: string): Promise<void> {
+	public async removeAdminRole(discordId: string): Promise<boolean> {
 		const idx = this.adminRoles.findIndex(r => r.discordId === discordId);
 		if (idx !== -1) {
 			const role = this.adminRoles[idx];
-			await getManager().remove(role);
+			await db.remove(role);
 			this.adminRoles.splice(idx, 1);
+			return true;
+		}
+		return false;
+	}
+
+	public async isMemberAdmin(member: GuildMember | string) {
+		const fetched = await this.fetchMember(member);
+		return fetched?.roles.cache.some(role => this.adminRoles.some(adminRole => adminRole.guild.discordId === this.discordId && adminRole.discordId === role.id)) ?? false;
+	}
+
+	private async fetchMember(member: GuildMember | string) {
+		if (member instanceof GuildMember) {
+			return await member.fetch();
+		} else {
+			return await Bot.instance.client.guilds.cache.get(this.discordId)?.members.fetch(member);
 		}
 	}
-};
+
+	public async save() {
+		await db.save(this);
+	}
+}
