@@ -1,4 +1,6 @@
-import { Entity, PrimaryColumn, Column, getManager, getRepository } from "typeorm";
+import { Entity, PrimaryColumn, Column } from "typeorm";
+
+import { db } from "../dataSource";
 import { IPlayerInfo } from "../model/IPlayerInfo";
 
 @Entity()
@@ -19,6 +21,9 @@ export class Player {
 	classId: number;
 
 	@Column()
+	gender: number;
+
+	@Column()
 	accountName: string;
 
 	@Column()
@@ -27,16 +32,23 @@ export class Player {
 	@Column({ nullable: true })
 	lastIpAddr: string;
 
-	public static async find(name: string): Promise<Player> {
-		return await getRepository(Player)
+	public static async findOne(name: string): Promise<Player | null> {
+		return await db.getRepository(Player)
 			.createQueryBuilder("player")
-			.where("player.name like :name", { name })
+			.where("LOWER(player.name) LIKE LOWER(:name)", { name: `${name}%` })
 			.getOne();
 	}
 
+	public static async find(name: string): Promise<Player[]> {
+		return await db.getRepository(Player)
+			.createQueryBuilder("player")
+			.where("LOWER(player.name) LIKE LOWER(:name)", { name: `${name}%` })
+			.getMany();
+	}
+
 	public static async save(data: IPlayerInfo): Promise<Player> {
-		let player = await getManager().findOne(Player, data.guid);
-		if (player === undefined) {
+		let player = await db.findOneBy(Player, { guid: data.guid });
+		if (!player) {
 			player = new Player();
 		}
 
@@ -45,11 +57,12 @@ export class Player {
 		player.level = data.level;
 		player.raceId = data.raceId;
 		player.classId = data.classId;
+		player.gender = data.gender;
 		player.accountName = data.accountName;
 		player.accountGuid = data.accountGuid;
 		player.lastIpAddr = data.lastIpAddr;
-		await getManager().save(player);
+		await db.save(player);
 
 		return player;
 	}
-};
+}
