@@ -2,4 +2,4 @@ export enum ChatMsg {
 	Say = 0x01,
 	Yell = 0x06,
 	Emote = 0x0A,
-};
+}
