@@ -4,7 +4,8 @@ export interface IPlayerInfo {
 	level: number;
 	raceId: number;
 	classId: number;
+	gender: number;
 	accountName: string;
 	accountGuid: number;
 	lastIpAddr: string;
-};
+}
